feat: restructure search generator
